+++ src/pudl/etl/__init__.py
@@ -82,6 +82,7 @@ core_module_groups = {
     "core_ferc714": [pudl.transform.ferc714],
     "core_gridpathratoolkit": [pudl.transform.gridpathratoolkit],
     "core_nrelatb": [pudl.transform.nrelatb],
+    "core_vceregen": [pudl.transform.vceregen],
 }
 
 out_module_groups = {