+++ docs/release_notes.rst
@@ -6,6 +6,19 @@ PUDL Release Notes
 v2024.X.x (2024-XX-XX)
 ---------------------------------------------------------------------------------------
 
+New Data
+^^^^^^^^
+
+Vibrant Clean Energy Renewable Generation Profiles
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+* Integrate the VCE hourly capacity factor data for solar pv, onshore, and offshore
+  wind from 2019 to 2024. The data in this table were produced by Vibrant Clean
+  Energy, and are licensed to the public under the Creative Commons Attribution
+  4.0 International license (CC-BY-4.0). Compare this data to the gridpathratoolkit
+  capacity factor data calculated at different granularities and using different
+  models.
+
+
 New Data Coverage
 ^^^^^^^^^^^^^^^^^